@@ -1,4 +1,4 @@
-import { CreateElement, FunctionalComponentOptions, RenderContext } from 'vue';
+import { defineComponent, h } from 'vue';
 
 export interface FileFieldValue {
   [propName: string]: any;
@@ -16,9 +16,9 @@ export interface FileProps {
   field: FileFieldValue | FileField;
 }
 
-export const File: FunctionalComponentOptions<FileProps> = {
+export const File = defineComponent({
   name: 'ScFileField',
-  functional: true,
+  inheritAttrs: false,
   props: {
     field: {
       type: Object,
@@ -29,31 +29,31 @@ export const File: FunctionalComponentOptions<FileProps> = {
   // Need to assign `any` return type because Vue type definitions are inaccurate.
   // The Vue type definitions set `render` to a return type of VNode and that's it.
   // However, it is possible to return null | string | VNode[] | VNodeChildrenArrayContents.
-  render(createElement: CreateElement, context: RenderContext): any {
-    const { field } = context.props;
+  render(): any {
+    const { field } = this.$props;
 
     /*
     File fields cannot be managed via the EE. We never output "editable."
   */
-    if (!field || (!field.value && !field.src)) {
+    if (!field || (!field.value && !(field as FileFieldValue).src)) {
       return null;
     }
 
     // handle link directly on field for forgetful devs
-    const file = field.src ? field : field.value;
+    const file = (field as FileFieldValue).src ? field : field.value;
     if (!file) {
       return null;
     }
 
-    if (context.data.scopedSlots && context.data.scopedSlots.default) {
-      return context.data.scopedSlots.default(file);
+    if (this.$slots && this.$slots.default) {
+      return this.$slots.default(file);
     }
 
     const linkText = file.title || file.displayName;
     // in functional components, context.data should be passed along to the
     // `createElement` function in order to retain attributes and events
     // https://vuejs.org/v2/guide/render-function.html#Passing-Attributes-and-Events-to-Child-Elements-Components
-    const data = { ...context.data, attrs: { ...context.data.attrs, href: file.src } };
-    return createElement('a', data, linkText);
+    const data = { ...this.$data, ...this.$attrs, href: file.src };
+    return h('a', data, linkText);
   },
-};
+});
