@@ -1,4 +1,4 @@
-import { CreateElement, FunctionalComponentOptions, RenderContext } from 'vue';
+import { h, defineComponent } from 'vue';
 
 export interface RichTextProps {
   /** The rich text field data. */
@@ -19,8 +19,7 @@ export interface RichTextProps {
   editable?: boolean;
 }
 
-export const RichText: FunctionalComponentOptions<RichTextProps> = {
-  functional: true,
+export const RichText = defineComponent({
   props: {
     field: { type: Object, required: true },
     tag: { type: String, default: 'div' },
@@ -29,8 +28,8 @@ export const RichText: FunctionalComponentOptions<RichTextProps> = {
   // Need to assign `any` return type because Vue type definitions are inaccurate.
   // The Vue type definitions set `render` to a return type of VNode and that's it.
   // However, it is possible to return null | string | VNode[] | VNodeChildrenArrayContents.
-  render(createElement: CreateElement, context: RenderContext): any {
-    const { field, tag, editable } = context.props;
+  render(): any {
+    const { field, tag, editable } = this.$props;
     if (!field || (!field.editable && !field.value)) {
       return null;
     }
@@ -39,10 +38,10 @@ export const RichText: FunctionalComponentOptions<RichTextProps> = {
     // `createElement` function in order to retain attributes and events
     // https://vuejs.org/v2/guide/render-function.html#Passing-Attributes-and-Events-to-Child-Elements-Components
     const data = {
-      ...context.data,
-      domProps: { innerHTML: field.editable && editable ? field.editable : field.value },
+      ...this.$data,
+      innerHTML: field.editable && editable ? field.editable : field.value,
     };
 
-    return createElement(tag || 'div', data);
+    return h(tag || 'div', data);
   },
-};
+});
