@@ -1,5 +1,4 @@
-import Vue, { PropOptions } from 'vue';
-import { ThisTypedComponentOptionsWithRecordProps } from 'vue/types/options';
+import { defineComponent, provide } from 'vue';
 import { ComponentFactory } from './sharedTypes';
 
 export interface SitecoreContextProps {
@@ -8,29 +7,18 @@ export interface SitecoreContextProps {
 
 // For an explanation of why `ThisTypedComponentOptionsWithRecordProps`
 // is used here, see the `Placeholder.ts` component file.
-export const SitecoreContext: ThisTypedComponentOptionsWithRecordProps<
-  Vue,
-  any,
-  any,
-  any,
-  SitecoreContextProps
-> = {
+export const SitecoreContext = defineComponent({
   props: {
     componentFactory: {
       // Hopefully this PR will negate having to cast the formatter prop
       // to PropOptions<any>: https://github.com/vuejs/vue/pull/6856
       type: Function,
       default: undefined,
-    } as PropOptions<any>,
+    },
   },
 
-  provide() {
-    return {
-      injectedComponentFactory: this.componentFactory,
-    };
+  setup(props, context) {
+    provide('injectedComponentFactory', props.componentFactory);
+    return () => context.slots && context.slots.default();
   },
-
-  render(): any {
-    return this.$slots && this.$slots.default[0];
-  },
-};
+});
