@@ -1,4 +1,4 @@
-import _Vue, { PluginObject } from 'vue';
+import { Plugin } from 'vue';
 import { ComponentFactory } from '../components/sharedTypes';
 import { providePlaceholders } from '../enhancers/providePlaceholders';
 
@@ -12,11 +12,11 @@ export interface SitecoreJssPlaceholderPluginOptions {
  * @param {typeof _Vue} Vue
  * @param {SitecoreJssPlaceholderPluginOptions} options
  */
-function install(Vue: typeof _Vue, options?: SitecoreJssPlaceholderPluginOptions) {
+function install(Vue, options?: SitecoreJssPlaceholderPluginOptions) {
   // eslint-disable-next-line no-param-reassign
-  Vue.prototype.$jss = {
+  Vue.config.globalProperties.$jss = {
     // there may be other JSS plugins installed
-    ...Vue.prototype.$jss,
+    ...Vue.config.globalProperties.$jss,
     componentFactory: options && options.componentFactory,
   };
 
@@ -27,12 +27,11 @@ function install(Vue: typeof _Vue, options?: SitecoreJssPlaceholderPluginOptions
   });
 }
 
-export const SitecoreJssPlaceholderPlugin: PluginObject<SitecoreJssPlaceholderPluginOptions> = {
+export const SitecoreJssPlaceholderPlugin: Plugin & SitecoreJssPlaceholderPluginOptions = {
   install,
-  key: 'SitecoreJssPlaceholderPlugin',
 };
 
-declare module 'vue/types/vue' {
+declare module 'vue/dist/vue' {
   export interface Vue {
     $jss: {
       componentFactory?: ComponentFactory;
