@@ -6,12 +6,11 @@
 </template>
 
 <script lang="ts">
-import Vue from 'vue';
 import { Placeholder } from '../../../components/Placeholder';
 
 // If VSCode shows "Default export of the module has or is using private name 'VueConstructor'.",
 // you can ignore it. The test and component should still compile and run.
-export default Vue.extend({
+export default {
   data() {
     return {
       arbitrary: { value: 'magical', withFunction: () => 'indeed' },
@@ -25,7 +24,7 @@ export default Vue.extend({
   components: {
     placeholder: Placeholder,
   },
-});
+};
 </script>
 
 <style>
