@@ -1,4 +1,6 @@
 declare module '*.vue' {
-  import Vue from 'vue';
-  export default Vue;
+  import { defineComponent } from 'vue';
+  const Component: ReturnType<typeof defineComponent>;
+
+  export default Component;
 }
