@@ -4,13 +4,14 @@
   <div id="root">
     <context-view :visible="contextViewVisible" />
     <!-- 'router-view' is a "global" component that is injected into the Vue component registry by vue-router. -->
-    <router-view v-if="!languageIsChanging" />
+    <router-view />
     <div v-if="languageIsChanging" class="alert alert-info">Language is changing...</div>
   </div>
 </template>
 
 <script>
 import ContextView from './ContextView';
+import { fetchDictionary } from './i18n';
 
 export default {
   name: 'AppRoot',
@@ -24,11 +25,14 @@ export default {
     changeAppLanguage(language) {
       // Changing languages is an async action, therefore the `languageIsChanging` property can be used
       // to show a loading/switching screen when language is being changed.
-      const i18n = this.$i18n.i18next;
-      if (i18n.language !== language) {
+      const i18n = this.$i18n;
+      if (i18n.locale !== language) {
         this.languageIsChanging = true;
 
-        i18n.changeLanguage(language, () => {
+        fetchDictionary(language).then((phrases) => {
+          i18n.setLocaleMessage(language, phrases);
+          i18n.locale = language;
+
           this.languageIsChanging = false;
         });
       }
@@ -41,6 +45,7 @@ export default {
     // Use Vue's provide/inject capabilities to "provide" functions to
     // any descendant component that want to use/"inject" the functions.
     return {
+      languageIsChanging: this.languageIsChanging,
       changeAppLanguage: this.changeAppLanguage,
       toggleContextView: this.toggleContextView,
     };
