@@ -17,9 +17,11 @@
       For instance, you may wish to render a button (<button />) instead of a link (<a />).
     -->
     <sc-file :field="fields.file" target="_blank">
-      <a slot-scope="file" target="_blank" v-bind:href="file.src"
-        >Custom link body ({{ file.title || file.displayName }})</a
-      >
+      <template v-slot="file">
+        <a target="_blank" v-bind:href="file.src"
+          >Custom link body ({{ file.title || file.displayName }})</a
+        >
+      </template>
     </sc-file>
   </styleguide-specimen>
 </template>
